% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{race}
\alias{race}
\title{Race two functions}
\usage{
race(a, b)
}
\arguments{
\item{a}{A process you want to compare the run time with process b}

\item{b}{The process to compare the run time against process a}
}
\value{
a length 1 double of the difference in seconds process a took compared to process b.
}
\description{
Race two functions
}
\examples{
race(rnorm(1e5),rnorm(1e7))
race(rnorm(1e7),rnorm(1e5))
}
